backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
environ==1.0
mssql-django==1.5
pyodbc==5.1.0
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
